Response userData is not freed here for Emscripten
When the user data is a pointer to the callback function, not a POST
body, it must not be freed, otherwise the app will crash/not function.
For now freeing it is handed to the user, which in this case, is
qrtclient. It's freed when it's a POST body, e.g. when logging in, and
no freeing code when GETting data from REST API.